requirements.txt
gradio
torch
Pillow
transformers
numpy
gradio-image-prompter
opencv-python